:goal_net: Correcion manejo de errores parte 1

+++ src/data/session/session.js
@@ -2,16 +2,52 @@ import { uriSession } from "./server";
 import axios from "axios";
 
 
+const METHOD_NOT_ALLOWED=405;
+const BAD_REQUEST=400;
+const CONFLICT=409;
+const NOT_ACCEPTABLE=406;
+const INTERNAL_ERROR = 500;
+const errorManager=(statusCode)=>{
+    switch (statusCode) {
+        case METHOD_NOT_ALLOWED:
+            return METHOD_NOT_ALLOWED + " No autorizado para hacer ejecutar esta petición";
+            break;
+        case BAD_REQUEST:
+            return BAD_REQUEST + " Falta algun campo";
+            break;
+        case CONFLICT:
+            return CONFLICT + " Se entró en conflico con la base de datos";
+            break;
+        case NOT_ACCEPTABLE:
+            return NOT_ACCEPTABLE + " Se ha superadó el tiempo permitido para hacer esta peticion";
+            break;
+        case INTERNAL_ERROR:
+            return INTERNAL_ERROR + " Error interno";
+            break;
+        default:
+            return CONFLICT + " Se entró en conflico con la base de datos";
+    }
+
+}
+
+
 /*
 CREATE
 */
 
-export const registerSchedule = async (schedule) => {
+export const registerSchedule = async (schedule, Type) => {
     try {
+
+        if(Type===2){
+            err = errorManager(METHOD_NOT_ALLOWED);
+            throw new Error(err)
+        }
+        
         const response = await axios.post(uriSession + '/schedule/create',schedule);
         return response.status;
     } catch (error) {
-        console.error(error);
+        err = errorManager(error.response.status);
+        throw new Error(err)
     }
 }
 
@@ -23,10 +59,16 @@ DELETE
 
 export const deleteSchedule = async (ChageState) => {
     try {
+        if(Type===2){
+            err = errorManager(METHOD_NOT_ALLOWED);
+            throw new Error(err)
+        }
+
         const response = await axios.delete(uriSession + '/schedule/delete',{ data: ChageState});
         return response.status;
     } catch (error) {
-        console.error(error);
+        err = errorManager(error.response.status);
+        throw new Error(err)
     }
 }
 
@@ -35,10 +77,15 @@ PUT
 */
 export const setAdate = async (ChageState) => {
     try {
+        if(Type===1){
+            err = errorManager(METHOD_NOT_ALLOWED);
+            throw new Error(err)
+        }
         const response = await axios.put(uriSession + '/set-a-date',ChageState);
         return response.status;
     } catch (error) {
-        console.error(error);
+        err = errorManager(error.response.status);
+        throw new Error(err)
     }
 }
 
@@ -47,8 +94,8 @@ export const calcelUser = async (ChageState) => {
         const response = await axios.put(uriSession + '/cancel/user',ChageState);
         return response.status;
     } catch (error) {
-        console.error(error);
-        console.error( error.response.status+" "+ error.response.statusCode );
+        err = errorManager(error.response.status)
+        throw new Error(err)
     }
 }
 
@@ -57,7 +104,8 @@ export const calcelCoach = async (ChageState) => {
         const response = await axios.put(uriSession + '/cancel/coach',ChageState);
         return response.status;
     } catch (error) {
-        console.error(error);
+        err = errorManager(error.response.status)
+        throw new Error(err)
     }
 }
 
@@ -69,7 +117,8 @@ export const getAllbyIdCoach = async (idCoach) => {
         const response = await axios.get(uriSession + '/get-by-idCoach/' + idCoach);
         return response.data;
     } catch (error) {
-        console.error(error);
+        err = errorManager(error.response.status)
+        throw new Error(err)
     }
 }
 export const getAllbyCoachCurrent = async (idCoach) => {
@@ -77,7 +126,8 @@ export const getAllbyCoachCurrent = async (idCoach) => {
         const response = await axios.get(uriSession + '/get-by-idCoach/Current/' + idCoach);
         return response.data;
     } catch (error) {
-        console.error(error);
+        err = errorManager(error.response.status)
+        throw new Error(err)
     }
 }
 
@@ -86,7 +136,8 @@ export const getAllbyIdUser = async (idUser) => {
         const response = await axios.get(uriSession + '/get-by-idUser/' + idUser);
         return response.data;
     } catch (error) {
-        console.error(error);
+        err = errorManager(error.response.status)
+        throw new Error(err)
     }
 }
 export const getAllbyUserCurrent = async (idUser) => {
@@ -94,6 +145,17 @@ export const getAllbyUserCurrent = async (idUser) => {
         const response = await axios.get(uriSession + '/get-by-idUser/Current/' + idUser);
         return response.data;
     } catch (error) {
-        console.error(error);
+        err = errorManager(error.response.status)
+        throw new Error(err)
+    }
+}
+export const getbyIdSchedule = async (idUser) => {
+    try {
+        const response = await axios.get(uriSession + '/get-by-idSchedule/' + idUser);
+        return response.data;
+    } catch (error) {
+        err = errorManager(error.response.status)
+        throw new Error(err)
+        
     }
 }
